Fixed Some TypeCasting issues

+++ app/src/main/java/com/karthihegde/readlist/retrofit/data/RetailPrice.kt
@@ -5,6 +5,6 @@ import kotlinx.parcelize.Parcelize
 
 @Parcelize
 data class RetailPrice(
-    val amountInMicros: Int,
+    val amountInMicros: Long,
     val currencyCode: String
 ) : Parcelable
